+++ lib/puppet-debugger/cli.rb
@@ -161,6 +161,10 @@ module PuppetDebugger
         exit 1 # this can sometimes causes tests to fail
       rescue PuppetDebugger::Exception::Error => e
         output = e.message.fatal
+      rescue RuntimeException => e
+        output = e.message.fatal
+        out_buffer.puts output
+        exit 1
       end
       unless output.empty?
         out_buffer.print ' => '